+++ mockserver-core/src/main/java/org/mockserver/mappers/JDKCertificateToMockServerX509Certificate.java
@@ -52,7 +52,9 @@ public class JDKCertificateToMockServerX509Certificate {
                     }
                 )
                 .collect(Collectors.toList());
-            httpRequest.withClientCertificateChain(x509Certificates);
+            if (!x509Certificates.isEmpty()) {
+                httpRequest.withClientCertificateChain(x509Certificates);
+            }
         }
         return httpRequest;
     }