+++ mockserver-core/src/main/java/org/mockserver/model/Headers.java
@@ -36,7 +36,6 @@ public class Headers extends KeysToMultiValues<Header, Headers> {
         return this;
     }
 
-    @SuppressWarnings("MethodDoesntCallSuperMethod")
     public Headers clone() {
         return new Headers(getMultimap());
     }