+++ mockserver-core/src/main/java/org/mockserver/serialization/serializers/request/HttpRequestDTOSerializer.java
@@ -48,6 +48,9 @@ public class HttpRequestDTOSerializer extends StdSerializer<HttpRequestDTO> {
         if (httpRequest.getSecure() != null) {
             jgen.writeBooleanField("secure", httpRequest.getSecure());
         }
+        if (httpRequest.getClientCertificateChain() != null && !httpRequest.getClientCertificateChain().isEmpty()) {
+            jgen.writeObjectField("clientCertificateChain", httpRequest.getClientCertificateChain());
+        }
         if (httpRequest.getSocketAddress() != null) {
             jgen.writeObjectField("socketAddress", httpRequest.getSocketAddress());
         }