+++ mockserver-core/src/main/java/org/mockserver/serialization/serializers/request/HttpRequestSerializer.java
@@ -48,6 +48,9 @@ public class HttpRequestSerializer extends StdSerializer<HttpRequest> {
         if (httpRequest.isSecure() != null) {
             jgen.writeBooleanField("secure", httpRequest.isSecure());
         }
+        if (httpRequest.getClientCertificateChain() != null && !httpRequest.getClientCertificateChain().isEmpty()) {
+            jgen.writeObjectField("clientCertificateChain", httpRequest.getClientCertificateChain());
+        }
         if (httpRequest.getSocketAddress() != null) {
             jgen.writeObjectField("socketAddress", httpRequest.getSocketAddress());
         }